Alpine 3.14 updated to gcc 10.3.
We ignore hadolint warning that ca-certificates should be versioned lock. For
this package the warning is misguided as the package in question contains
certificates that have an expiry date. Thus 1) there will only be one correct
version at any time in the repo and 2) not locking the version makes the build
more repeatable.