Publish a GitHub Action for gitops-promotion.
Because GitHub only allows a single action in each repo, we add
a custom endpoint that allows the action to pass the command as
an environment var along the rest of the arguments.

The current version of gitops-promotion has some quirk in how it
behaves that means it needs access to the Git history. This is
problematic in that the default GitHub checkout Action defaults
to --depth=1, prompting gitops-promotion to fail with "object not
found - no match for id (SHA)" when pushing the promotion branch.
Despite enthusiastic testing with plumbing commands, I cannot
reproduce this state locally, and so this limitation will have to
stand for now.

Signed-off-by: Anders Qvist <anders.qvist@xenit.se>